==> 1_Data_preprocessing.ipynb <==
## Table of Content
1. Basic Imports
2. Meta-Data 
3. Credits
4. keywords
5. Merge and Saving

## 1.Basic Imports

import pandas as pd
import numpy as np
import re
import ast

## 2. Metadata

meta=pd.read_csv("Dataset/movies_metadata.csv")
meta.head()

meta.columns

meta = meta[meta['id'].notna()]
meta = meta[meta['title'].notna()]
meta.reset_index(inplace=True,drop=True)
meta.head()

meta.drop(['adult',
       'homepage', 'poster_path','budget','revenue','tagline','overview','imdb_id',
       'production_countries', 'runtime','status','original_language',
       'spoken_languages', 'video','original_title'],axis=1,inplace=True)

meta.head()

meta['year'] = pd.to_datetime(meta['release_date'], errors='coerce').apply(lambda x: str(x).split('-')[0] if x != np.nan else np.nan)
meta.drop(['release_date'],axis=1,inplace=True)
meta.head()

### Genres, Companies and Collection are in json format
### Extracting the required info into the simpler format

def generate_genre(data):
    genres=dict()

    for i in range(len(data)):
        x=data[i]
        l=list()
        for a in x:     
            id=a.get('id')
            name=a.get('name')

            genres[id]=name
            l.append(name)

        data[i]=", ".join(l)

    return genres
        

def generate_production_companies(data):
    comp=dict()

    for i in range(len(data)):
        x=data[i]
        l=list()
        for a in x:
            id=a.get('id')
            name=a.get('name')

            comp[id]=name
            l.append(name)

        data[i]=", ".join(l)

    return comp

def generate_collection(data):
    coll=dict()

    for i in range(len(data)):
        a=data[i]

        if a=='':
            continue
        id=a.get('id')
        name=a.get('name')

        coll[id]=name

        data[i]=name

    return coll

def perfectEval(anonstring):
    try:
        ev = ast.literal_eval(anonstring)
        return ev
    except ValueError:
        
        corrected = "\'" + "\'"
        ev = ast.literal_eval(corrected)
        return ev

meta['genres'] = meta['genres'].map(lambda x: perfectEval(x))
meta['production_companies'] = meta['production_companies'].map(lambda x: perfectEval(x))
meta['belongs_to_collection'] = meta['belongs_to_collection'].map(lambda x: perfectEval(x))

meta.head()

genres_dict=generate_genre(meta['genres'])
company_dict=generate_production_companies(meta['production_companies'])
collection_dict=generate_collection(meta['belongs_to_collection'])
meta.head()

genres_dict

meta.shape

## 3. Credits 

credits=pd.read_csv('Dataset/credits.csv')
credits.head()

credits.shape

### Cast and Crew are in json format
### Extracting the required info into the simpler format

def generate_cast(data):
    casts=dict()

    for i in range(len(data)):
        x=data[i]
        l=list()

        for a in x:
            id=a.get('cast_id')
            name=a.get('name')

            casts[id]=name
            l.append(name)
        
        data[i]=', '.join(l[:3])
    return casts


def generate_dir(data):
    dirct=dict()

    for i in range(len(data)):
        x=data[i]
        l=list()

        for a in x:
            if a.get('job')!='Director':
                continue
            
            id=a.get('cast_id')
            name=a.get('name')

            dirct[id]=name
            l.append(name)
        
        data[i]=", ".join(l)
    return dirct


credits['cast']=credits['cast'].map(lambda x : perfectEval(x))
credits['crew']=credits['crew'].map(lambda x : perfectEval(x))


dir_dict=generate_dir(credits['crew'])
cast_dict=generate_cast(credits['cast'],)
credits.columns=['cast','director','id']

credits.head()

## 4. Keywords

key=pd.read_csv('Dataset/keywords.csv')
key.head()

key.shape

### Keywords is in json format
### Extracting the required info into the simpler format

def generate_keywords(data):
    keys=dict()

    for i in range(len(data)):
        x=data[i]
        l=list()

        for a in x:
            id=a.get('id')
            name=a.get('name')

            keys[id]=name
            l.append(name)
        data[i]=', '.join(l)
    return data,keys

key['keywords']=key['keywords'].map(lambda x : perfectEval(x))

key['keywords'],key_dict=generate_keywords(key['keywords'])

key.head()

**Keeping only the top-repeating keywords**

keywords=list()

for row in key.keywords:
    x=row.split(', ')
    for name in x:
        if name=='':
            continue
        keywords.append(name)
from collections import Counter
c = Counter(keywords)

top_keywords=[x[0] for x in c.most_common(3000)]

top_keywords[:10]

for i in range(len(key.id)):
    row=key['keywords'][i].split(', ')
    y=[x for x in row if x in top_keywords]
    
    key['keywords'][i]=', '.join(y)

key.head()

## 5. Merge and Save

meta['id']=meta['id'].astype(int)
credits['id']=credits['id'].astype(int)
key['id']=key['id'].astype(int)

final_metadata=meta.merge(credits,on="id").merge(key,on='id')
final_metadata.head()

final_metadata= final_metadata[['id', 'title','year', 'genres','cast','director', 'belongs_to_collection', 'production_companies','vote_count','vote_average','popularity','keywords']]
final_metadata.rename(columns={'belongs_to_collection':'collection','id':'movieId'},inplace=True)
final_metadata.head()

final_metadata = final_metadata[final_metadata['movieId'].notna()]
final_metadata.fillna('',inplace=True)
final_metadata.drop_duplicates(inplace=True)

final_metadata.reset_index(drop=True,inplace=True)

final_metadata.to_csv('Dataset/prepared/final_metadata.csv',index=False)

import json

with open('Dataset/prepared/company_index.json', 'w') as fp:
    json.dump(company_dict, fp)


with open('Dataset/prepared/collection_index.json', 'w') as fp:
    json.dump(collection_dict, fp)


with open('Dataset/prepared/genre_index.json', 'w') as fp:
    json.dump(genres_dict, fp)


with open('Dataset/prepared/cast_index.json', 'w') as fp:
    json.dump(cast_dict, fp)


with open('Dataset/prepared/dir_index.json', 'w') as fp:
    json.dump(dir_dict, fp)


with open('Dataset/prepared/key_index.json', 'w') as fp:
    json.dump(key_dict, fp)
