==> 2_Simple-Recommender.ipynb <==
# Table of Content
1. Introduction and Import
2. Top Movies of Different Genres
3. Feature Extraction, User Likings and Implementation

import pandas as pd

metadata=pd.read_csv('Dataset/prepared/final_metadata.csv')
metadata.head()

# Based on Different Genres


df1=metadata[['movieId','title','genres','popularity']]
df1.fillna('',inplace=True)
df1.head()

## Top Drama Movies

df1[df1['genres'].str.contains("Drama")].sort_values('popularity',ascending=False).head(10)

## Top Animation Movies

df1[df1['genres'].str.contains("Animation")].sort_values('popularity',ascending=False).head(10)

## Top Action movies

df1[df1['genres'].str.contains("Action")].sort_values('popularity',ascending=False).head(10)

## Top Comedy Movies

df1[df1['genres'].str.contains("Comedy")].sort_values('popularity',ascending=False).head(10)
