==> 5_Cold_Start.ipynb <==
## Table of Contents
1. Imports and Introduction
2. Split Loc. and Missing-Handling
3. Implementation

## Imports and Introduction

import pandas as pd
import numpy as np

user_details=pd.read_csv('Dataset/Users.csv')
user_details.head()

rat=pd.read_csv('Dataset/ratings.csv')
rat.head()

## Split Location and Handling Missing

user_details['Location']=user_details['Location'].str.split(', ')
user_details.head()

def return_area(arr):
    if len(arr)<3:
        return ''
    return arr[0]

def return_town(arr):
    if len(arr)<3:
        return ''
    return arr[1]

def return_country(arr):
    if len(arr)<3:
        return ''
    return arr[2]

user_details['area']=user_details['Location'].apply(return_area)
user_details['town']=user_details['Location'].apply(return_town)
user_details['country']=user_details['Location'].apply(return_country)

user_details.head()

a=user_details['Age'].mean()
user_details['Age'].fillna(a,inplace=True)
user_details.head()

user_details.isna().sum()

## Implementation



def new_user_recommend(age,loc,users,ratings):
    
#     spliting the location into area, town, country
    loc=loc.split(', ')
    
#     selecting the users from same area
    users_new=users[users.area==loc[0]]

#     if there are only few users in that area selecting from same town
    if len(users_new)<=1000:
        users_new=users[users.town==loc[1]]
    
#     if the users are still less taking the users from same country
    if len(users_new)<=1000:
        users_new=users[users.country==loc[2]]
    
        
#     selecting the top 50 users nearest to that age of new user 
    new_list=users_new.iloc[(users_new['Age']-age).abs().argsort()[:50]]
    
#     extracting the user-id of those users
    l=list(new_list['User-ID'])
    
#     selecting the rating of only those users
    mask = rat.userId.apply(lambda x: x in l)
    selected_ratings= ratings[mask]
    
#     selecting the top rated movies by those users
    top_ratings=selected_ratings.sort_values('rating',ascending=False)
    
#     list of movie-id's
    movie_list= list(top_ratings.movieId)[:15]
    
    return movie_list
    



new_user_recommend(26,'nyc, new york, usa',user_details,rat)
