==> 6_User_Liking_Based.ipynb <==
# Table of Content
1. Introduction and Import
2. Feature Extraction
3. User Likings and Implementation

import pandas as pd

metadata=pd.read_csv('Dataset/prepared/final_metadata.csv')
metadata.head()

# 2. Features Extraction

movies=metadata[['movieId','genres','cast','keywords']]

movies.head()

# creating a feature column consisting of genres, cast and keywords
features=movies.copy()
features['features']=''
features.fillna(' ',inplace=True)
features['features']=features['genres']+', '+features['cast']+features['keywords']
features.drop(['genres','cast','keywords'],axis=1,inplace=True)
features['features']=features['features'].str.split(', ')
features.head()

rating=pd.read_csv('Dataset/ratings.csv')
rating.drop(['timestamp'],axis=1,inplace=True)
rating.columns=['userId','movieId','rating']

rating=rating.head(50000)
rating.head()

# merging the ratings with the features
merged=pd.merge(rating,features,on='movieId')
merged.head()

# calculating the average rating given by the user to the specific feature
# if user gives a rating to a certain movie, then the rating to all the features of that movie is considered to be same
# after that all the average rating for that specific feature is listed in th below df

likings_sum = merged.explode('features').pivot_table(index='userId', columns='features', values='rating', aggfunc='sum').fillna(0, downcast='infer')
likings_count = merged.explode('features').pivot_table(index='userId', columns='features', values='rating', aggfunc='count').fillna(0, downcast='infer')

likings_count['total']=likings_count.sum(axis=1)
likings_count.head()

likings_sum.head()

## 3. Implementation

import numpy as np

# this function calculates the average rating given by the user(specified) to all the features 
#  of that movie(specified) 
#  after that averages of that averages is returned as expected to be the rating of that user for that movie

def predict(user,movie,likings_count,likings_sum,features):
    
    if user not in list(likings_count.index):
        print('User not found')
        return 0
    if movie not in list(features.movieId):
        print('movie not found')
        return 0
    
#     extracting the features of the movie
    fet=features[features.movieId==movie].features
    fet=fet.values[0]

#     calculating the average rating given by the user to the features of the movie

    avg=[]
    for i in fet:
        c=likings_count.loc[user][i]
        s=likings_sum.loc[user][i]
        if c==0:
            avg.append(0)
        else:
            avg.append(s/c)

#     calculating the average of the averages
    avg=np.array(avg)
    avg=avg[avg!=0]
    if len(avg)==0:
        return 0

    return avg.mean()

user=11
movie=110

# for the user 11 , the predicted rating for the movie 110 is 3.50521448
predict(user,movie,likings_count,likings_sum,features)

likings_count.to_csv('Dataset/prepared/user_likings_count.csv')
likings_sum.to_csv('Dataset/prepared/user_likings_sum.csv')
features.to_csv('Dataset/prepared/movie_features.csv',index=False)
