==> 7_Clustering.ipynb <==

# Importing the libraries
import pandas as pd
from sklearn.cluster import KMeans

# Importing the dataset
rating=pd.read_csv('Dataset/ratings.csv')
rating=rating.head(100000)
rating.head()

# Importing the dataset-feature

fet=pd.read_csv('Dataset/prepared/user_likings.csv')
fet.head()



users=list(fet.userId)
fet.index=users
fet.drop(['userId'],inplace=True,axis=1)
fet.head()


# Clustering
kmeans = KMeans(
     init="random",
        n_clusters=25,
        n_init=10,
        max_iter=300,
        random_state=42
    )
    
kmeans.fit(fet)


labels=list(kmeans.labels_)
labels[:5]

# Creating a dictionary of users and their cluster
user_cluster=dict(zip(users,labels))

# Creating a dictionary of cluster and their users
cluster_user={}
for i in range(25):
    cluster_user[i]=[]

for i in user_cluster:
    cluster_user[user_cluster[i]].append(i)

    


def cluster_rat(rating,user_cluster,cluster_no,userid):
    '''
    rating: dataframe of ratings
    user_cluster: dictionary of user and their cluster
    cluster_no: cluster number
    userid: user id
    '''
    # Finding the users in the cluster
    users=cluster_user[cluster_no]
    # Finding the movies rated by the users in the cluster
    movies=[]
    for i in users:
        movies.extend(list(rating[rating.userId==i].movieId))
    # Finding the movies rated by the user
    movies_user=list(rating[rating.userId==userid].movieId)
    # Finding the movies rated by the user and the users in the cluster
    movies_common=list(set(movies).intersection(movies_user))
    # Finding the movies not rated by the user but rated by the users in the cluster
    movies_not_common=list(set(movies).difference(movies_user))
    # Finding the ratings of the movies rated by the users in the cluster
    rating_common=rating[rating.movieId.isin(movies_common)]
    # Finding the ratings of the movies not rated by the user but rated by the users in the cluster
    rating_not_common=rating[rating.movieId.isin(movies_not_common)]
    # Finding the average rating of the movies rated by the users in the cluster
    rating_common=rating_common.groupby('movieId').mean()
    # Finding the average rating of the movies not rated by the user but rated by the users in the cluster
    rating_not_common=rating_not_common.groupby('movieId').mean()
    # Finding the movies which are rated by the users in the cluster and the user
    movies_common=list(rating_common.index)
    # Finding the movies which are not rated by the users in the cluster and the user
    movies_not_common=list(rating_not_common.index)
    # Finding the average rating of the movies rated by the users in the cluster and the user
    rating_common=rating_common['rating']
    # Finding the average rating of the movies not rated by the users in the cluster and the user
    rating_not_common=rating_not_common['rating']
    # Finding the average rating of the movies rated by the users in the cluster and the user
    rating_common=pd.DataFrame(rating_common)
    # Finding the average rating of the movies not rated by the users in the cluster and the user
    rating_not_common=pd.DataFrame(rating_not_common)
    # Finding the average
    
cluster_rat(rating,user_cluster,1,1221)
